api_ingestor/requirements.txt
Flask
requests
beautifulsoup4
psycopg2-binary
celery[redis]
celery[beat]
redis
flower
pandas
